refactor(desktop): split state operations by ownership
