+++ apps/desktop/src/pages/agents/use-agent-studio-build-worktree-refresh.ts
@@ -1,5 +1,8 @@
 import { useEffect, useRef } from "react";
-import { isReadOnlyShellCommand, isSafeReadToolName } from "@/state/operations/permission-policy";
+import {
+  isReadOnlyShellCommand,
+  isSafeReadToolName,
+} from "@/state/operations/shared/permission-policy";
 import type { AgentSessionState } from "@/types/agent-orchestrator";
 
 type UseAgentStudioBuildWorktreeRefreshArgs = {
