+++ apps/desktop/src/pages/agents/use-agent-studio-rebase-conflict-resolution.ts
@@ -8,7 +8,7 @@ import {
 } from "@/features/git-conflict-resolution";
 import type { AgentSessionState } from "@/types/agent-orchestrator";
 import type { AgentStateContextValue } from "@/types/state-slices";
-import { loadEffectivePromptOverrides } from "../../state/operations/prompt-overrides";
+import { loadEffectivePromptOverrides } from "../../state/operations/shared/prompt-overrides";
 import type { AgentStudioQueryUpdate } from "./agent-studio-navigation";
 import {
   resolveAgentStudioBuilderSessionForTask,
