+++ apps/desktop/src/state/operations/agent-orchestrator/agent-tool-messages.ts
@@ -1,4 +1,4 @@
-import type { AgentChatMessage } from "../../types/agent-orchestrator";
+import type { AgentChatMessage } from "../../../types/agent-orchestrator";
 
 type ToolStatus = "pending" | "running" | "completed" | "error";
 type ToolCompletionOutcome = "completed" | "error";
