+++ apps/desktop/src/state/operations/agent-orchestrator/events/session-tool-parts.ts
@@ -1,6 +1,6 @@
 import { isOdtWorkflowMutationToolName } from "@openducktor/core";
 import type { AgentChatMessageMeta, AgentSessionState } from "@/types/agent-orchestrator";
-import { formatToolContent, isTodoToolName } from "../../agent-tool-messages";
+import { formatToolContent, isTodoToolName } from "../agent-tool-messages";
 import { runOrchestratorSideEffect } from "../support/async-side-effects";
 import { upsertMessage } from "../support/messages";
 import {
