+++ apps/desktop/src/state/operations/index.ts
@@ -1,8 +1,8 @@
-export { createHostRuntimeCatalogOperations } from "./runtime-catalog";
-export { useAgentOrchestratorOperations } from "./use-agent-orchestrator-operations";
-export { useChecks } from "./use-checks";
-export { useDelegationOperations } from "./use-delegation-operations";
-export { useRepoSettingsOperations } from "./use-repo-settings-operations";
-export { useSpecOperations } from "./use-spec-operations";
-export { useTaskOperations } from "./use-task-operations";
-export { useWorkspaceOperations } from "./use-workspace-operations";
+export { useAgentOrchestratorOperations } from "./agent-orchestrator/use-agent-orchestrator-operations";
+export { createHostRuntimeCatalogOperations } from "./shared/runtime-catalog";
+export { useDelegationOperations } from "./tasks/use-delegation-operations";
+export { useSpecOperations } from "./tasks/use-spec-operations";
+export { useTaskOperations } from "./tasks/use-task-operations";
+export { useChecks } from "./workspace/use-checks";
+export { useRepoSettingsOperations } from "./workspace/use-repo-settings-operations";
+export { useWorkspaceOperations } from "./workspace/use-workspace-operations";
