+++ apps/desktop/src/state/operations/shared/host.ts
@@ -1,0 +1,1 @@
+export { hostClient as host } from "@/lib/host-client";
