+++ apps/desktop/src/state/operations/workspace/workspace-operations-model.test.ts
