+++ apps/desktop/src/state/operations/workspace/workspace-operations-model.ts
